Fixed input bug and optimized the code
